+++ backend/src/main/java/tw/waterballsa/gaas/unoflip/usecase/GameJoinUseCase.java
@@ -1,4 +1,4 @@
-package tw.waterballsa.gaas.unoflip.service;
+package tw.waterballsa.gaas.unoflip.usecase;
 
 import org.springframework.stereotype.Service;
 import tw.waterballsa.gaas.unoflip.domain.UnoFlipGame;
@@ -6,10 +6,10 @@ import tw.waterballsa.gaas.unoflip.repository.GameRepo;
 import tw.waterballsa.gaas.unoflip.vo.GameJoinResult;
 
 @Service
-public class GameUseCase {
+public class GameJoinUseCase {
     private final GameRepo gameRepo;
 
-    public GameUseCase(GameRepo gameRepo) {
+    public GameJoinUseCase(GameRepo gameRepo) {
         this.gameRepo = gameRepo;
     }
